session: migrate test-infra to testify for TestGlobalAndLocalTxn
