+++ session/session_legacy_test.go
@@ -3002,181 +3002,6 @@ func (s *testSessionSerialSuite) TestSetTxnScope(c *C) {
 	failpoint.Disable("tikvclient/injectTxnScope")
 }
 
-func (s *testSessionSerialSuite) TestGlobalAndLocalTxn(c *C) {
-	// Because the PD config of check_dev_2 test is not compatible with local/global txn yet,
-	// so we will skip this test for now.
-	if *withTiKV {
-		return
-	}
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("set global tidb_enable_local_txn = on;")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("drop placement policy if exists p1")
-	tk.MustExec("drop placement policy if exists p2")
-	tk.MustExec("create placement policy p1 leader_constraints='[+zone=dc-1]'")
-	tk.MustExec("create placement policy p2 leader_constraints='[+zone=dc-2]'")
-	tk.MustExec(`create table t1 (c int)
-PARTITION BY RANGE (c) (
-	PARTITION p0 VALUES LESS THAN (100) placement policy p1,
-	PARTITION p1 VALUES LESS THAN (200) placement policy p2
-);`)
-	defer func() {
-		tk.MustExec("drop table if exists t1")
-		tk.MustExec("drop placement policy if exists p1")
-		tk.MustExec("drop placement policy if exists p2")
-	}()
-
-	// set txn_scope to global
-	tk.MustExec(fmt.Sprintf("set @@session.txn_scope = '%s';", kv.GlobalTxnScope))
-	result := tk.MustQuery("select @@txn_scope;")
-	result.Check(testkit.Rows(kv.GlobalTxnScope))
-
-	// test global txn auto commit
-	tk.MustExec("insert into t1 (c) values (1)") // write dc-1 with global scope
-	result = tk.MustQuery("select * from t1")    // read dc-1 and dc-2 with global scope
-	c.Assert(len(result.Rows()), Equals, 1)
-
-	// begin and commit with global txn scope
-	tk.MustExec("begin")
-	txn, err := tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	c.Assert(tk.Se.GetSessionVars().TxnCtx.TxnScope, Equals, kv.GlobalTxnScope)
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("insert into t1 (c) values (1)") // write dc-1 with global scope
-	result = tk.MustQuery("select * from t1")    // read dc-1 and dc-2 with global scope
-	c.Assert(len(result.Rows()), Equals, 2)
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("commit")
-	result = tk.MustQuery("select * from t1")
-	c.Assert(len(result.Rows()), Equals, 2)
-
-	// begin and rollback with global txn scope
-	tk.MustExec("begin")
-	txn, err = tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	c.Assert(tk.Se.GetSessionVars().TxnCtx.TxnScope, Equals, kv.GlobalTxnScope)
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("insert into t1 (c) values (101)") // write dc-2 with global scope
-	result = tk.MustQuery("select * from t1")      // read dc-1 and dc-2 with global scope
-	c.Assert(len(result.Rows()), Equals, 3)
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("rollback")
-	result = tk.MustQuery("select * from t1")
-	c.Assert(len(result.Rows()), Equals, 2)
-
-	timeBeforeWriting := time.Now()
-	tk.MustExec("insert into t1 (c) values (101)") // write dc-2 with global scope
-	result = tk.MustQuery("select * from t1")      // read dc-1 and dc-2 with global scope
-	c.Assert(len(result.Rows()), Equals, 3)
-
-	failpoint.Enable("tikvclient/injectTxnScope", `return("dc-1")`)
-	defer failpoint.Disable("tikvclient/injectTxnScope")
-	// set txn_scope to local
-	tk.MustExec("set @@session.txn_scope = 'local';")
-	result = tk.MustQuery("select @@txn_scope;")
-	result.Check(testkit.Rows("local"))
-
-	// test local txn auto commit
-	tk.MustExec("insert into t1 (c) values (1)")          // write dc-1 with dc-1 scope
-	result = tk.MustQuery("select * from t1 where c = 1") // point get dc-1 with dc-1 scope
-	c.Assert(len(result.Rows()), Equals, 3)
-	result = tk.MustQuery("select * from t1 where c < 100") // read dc-1 with dc-1 scope
-	c.Assert(len(result.Rows()), Equals, 3)
-
-	// begin and commit with dc-1 txn scope
-	tk.MustExec("begin")
-	txn, err = tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	c.Assert(tk.Se.GetSessionVars().CheckAndGetTxnScope(), Equals, "dc-1")
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("insert into t1 (c) values (1)")            // write dc-1 with dc-1 scope
-	result = tk.MustQuery("select * from t1 where c < 100") // read dc-1 with dc-1 scope
-	c.Assert(len(result.Rows()), Equals, 4)
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("commit")
-	result = tk.MustQuery("select * from t1 where c < 100")
-	c.Assert(len(result.Rows()), Equals, 4)
-
-	// begin and rollback with dc-1 txn scope
-	tk.MustExec("begin")
-	txn, err = tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	c.Assert(tk.Se.GetSessionVars().CheckAndGetTxnScope(), Equals, "dc-1")
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("insert into t1 (c) values (1)")            // write dc-1 with dc-1 scope
-	result = tk.MustQuery("select * from t1 where c < 100") // read dc-1 with dc-1 scope
-	c.Assert(len(result.Rows()), Equals, 5)
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("rollback")
-	result = tk.MustQuery("select * from t1 where c < 100")
-	c.Assert(len(result.Rows()), Equals, 4)
-
-	// test wrong scope local txn auto commit
-	_, err = tk.Exec("insert into t1 (c) values (101)") // write dc-2 with dc-1 scope
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*out of txn_scope.*")
-	err = tk.ExecToErr("select * from t1 where c = 101") // point get dc-2 with dc-1 scope
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*can not be read by.*")
-	err = tk.ExecToErr("select * from t1 where c > 100") // read dc-2 with dc-1 scope
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*can not be read by.*")
-	tk.MustExec("begin")
-	err = tk.ExecToErr("select * from t1 where c = 101") // point get dc-2 with dc-1 scope
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*can not be read by.*")
-	err = tk.ExecToErr("select * from t1 where c > 100") // read dc-2 with dc-1 scope
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*can not be read by.*")
-	tk.MustExec("commit")
-
-	// begin and commit reading & writing the data in dc-2 with dc-1 txn scope
-	tk.MustExec("begin")
-	txn, err = tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	c.Assert(tk.Se.GetSessionVars().CheckAndGetTxnScope(), Equals, "dc-1")
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("insert into t1 (c) values (101)")       // write dc-2 with dc-1 scope
-	err = tk.ExecToErr("select * from t1 where c > 100") // read dc-2 with dc-1 scope
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*can not be read by.*")
-	tk.MustExec("insert into t1 (c) values (99)")           // write dc-1 with dc-1 scope
-	result = tk.MustQuery("select * from t1 where c < 100") // read dc-1 with dc-1 scope
-	c.Assert(len(result.Rows()), Equals, 5)
-	c.Assert(txn.Valid(), IsTrue)
-	_, err = tk.Exec("commit")
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*out of txn_scope.*")
-	// Won't read the value 99 because the previous commit failed
-	result = tk.MustQuery("select * from t1 where c < 100") // read dc-1 with dc-1 scope
-	c.Assert(len(result.Rows()), Equals, 4)
-
-	// Stale Read will ignore the cross-dc txn scope.
-	c.Assert(tk.Se.GetSessionVars().CheckAndGetTxnScope(), Equals, "dc-1")
-	result = tk.MustQuery("select @@txn_scope;")
-	result.Check(testkit.Rows("local"))
-	err = tk.ExecToErr("select * from t1 where c > 100") // read dc-2 with dc-1 scope
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Matches, ".*can not be read by.*")
-	// Read dc-2 with Stale Read (in dc-1 scope)
-	timestamp := timeBeforeWriting.Format(time.RFC3339Nano)
-	// TODO: check the result of Stale Read when we figure out how to make the time precision more accurate.
-	tk.MustExec(fmt.Sprintf("select * from t1 AS OF TIMESTAMP '%s' where c = 101", timestamp))
-	tk.MustExec(fmt.Sprintf("select * from t1 AS OF TIMESTAMP '%s' where c > 100", timestamp))
-	tk.MustExec(fmt.Sprintf("START TRANSACTION READ ONLY AS OF TIMESTAMP '%s'", timestamp))
-	tk.MustExec("select * from t1 where c = 101")
-	tk.MustExec("select * from t1 where c > 100")
-	tk.MustExec("commit")
-	tk.MustExec("set @@tidb_replica_read='closest-replicas'")
-	tk.MustExec(fmt.Sprintf("select * from t1 AS OF TIMESTAMP '%s' where c > 100", timestamp))
-	tk.MustExec(fmt.Sprintf("START TRANSACTION READ ONLY AS OF TIMESTAMP '%s'", timestamp))
-	tk.MustExec("select * from t1 where c = 101")
-	tk.MustExec("select * from t1 where c > 100")
-	tk.MustExec("commit")
-
-	tk.MustExec("set global tidb_enable_local_txn = off;")
-}
-
 func (s *testSessionSuite2) TestSetEnableRateLimitAction(c *C) {
 	tk := testkit.NewTestKitWithInit(c, s.store)
 	// assert default value
